requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
numpy>=1.26.0
pandas>=2.1.0
mysql-connector-python>=8.2.0
scikit-learn>=1.3.0
tensorflow>=2.16.0 